refactor(ui): extract shared graph canvas into graph-canvas.jsx
Moves GR_NODE_SIZE/GR_Canvas/GR_NodeBox/GR_EdgePath/GR_SingleEdge and the
fan-out edge helpers verbatim out of graphs.jsx into a shared module so the
session-detail run view can reuse the same canvas. No behavior change; the
graph editor references them by name via the concatenated bundle scope.
