+++ ui/index.html
@@ -68,6 +68,7 @@
 <script type="text/babel" src="components/shared/bottom-sheet.jsx"></script>
 <script type="text/babel" src="components/shared/mobile-tabs.jsx"></script>
 <script type="text/babel" src="components/shared/floating-action.jsx"></script>
+<script type="text/babel" src="components/graph-canvas.jsx"></script>
 <script type="text/babel" src="components/mock-data.jsx"></script>
 <script type="text/babel" src="components/predicate-builder.jsx"></script>
